Revert "rename css file more concisely"
This reverts commit e8334c0b279a0bce695d702b12edb556096ef67b.
